Potrosnja iz inta u float

+++ models/IzvestajPoMesecu.py
@@ -3,7 +3,7 @@ class IzvestajMesec():
         self.items = items
 
 class IzvestajMesecItem():
-    def __init__(self, brojilo : int, korisnik : str, adresa : str, grad : str, potrosnja : int):
+    def __init__(self, brojilo : int, korisnik : str, adresa : str, grad : str, potrosnja : float):
         self.brojilo = brojilo
         self.korisnik = korisnik
         self.grad = grad
